requirements.txt
schedule~=1.2.2
python-telegram-bot~=21.10
requests~=2.32.3
pyyaml~=6.0.2